Fixed code scanning warnings

+++ src/main/java/nu/sitia/loggenerator/util/Ipv4.java
@@ -33,7 +33,7 @@ public class Ipv4 {
         for (int i = 0; i < ipAddressInArray.length; i++) {
             int power = 3 - i;
             int ip = Integer.parseInt(ipAddressInArray[i]);
-            result += ip * Math.pow(256, power);
+            result += (long) (ip * Math.pow(256, power));
         }
         return result;
     }
